attached: refactor user_v2 handler
