+++ embed/oko_attached/src/window_msgs/oauth_info_pass/handlers/ed25519_keygen.ts
@@ -1,0 +1,620 @@
+import type { AuthType } from "@oko-wallet/oko-types/auth";
+import type { KeyShareNodeMetaWithNodeStatusInfo } from "@oko-wallet/oko-types/tss";
+import { teddsaKeyShareToHex } from "@oko-wallet/oko-types/user_key_share";
+import type { Result } from "@oko-wallet/stdlib-js";
+import { type OAuthSignInError } from "@oko-wallet/oko-sdk-core";
+import {
+  serializeKeyPackage,
+  serializePublicKeyPackage,
+} from "@oko-wallet/teddsa-hooks";
+import { reqKeygenEd25519 } from "@oko-wallet/teddsa-api-lib";
+
+import {
+  makeAuthorizedOkoApiRequest,
+  signInV2,
+  TSS_V2_ENDPOINT,
+} from "@oko-wallet-attached/requests/oko_api";
+import { combineUserShares } from "@oko-wallet-attached/crypto/combine";
+import type { UserSignInResultV2 } from "@oko-wallet-attached/window_msgs/types";
+import { runExpandShares } from "@oko-wallet-attached/crypto/reshare";
+import { expandAndSendReshareV2 } from "@oko-wallet-attached/crypto/reshare_v2";
+import {
+  requestKeySharesV2,
+  requestKeySharesV2WithReshareInfo,
+  registerKeyShareEd25519V2,
+  reshareKeySharesV2,
+  reshareRegisterV2,
+} from "@oko-wallet-attached/requests/ks_node_v2";
+import {
+  commitAll,
+  createOkoApiCommitRevealParams,
+  createKsnCommitRevealParams,
+  type KsnCommitTarget,
+} from "@oko-wallet-attached/crypto/commit_reveal";
+import type { ReshareRequestV2 } from "@oko-wallet/oko-types/user";
+import {
+  decodeSecp256k1SharesByNode,
+  encodePoint256ToKeyShareString,
+} from "@oko-wallet-attached/crypto/key_share_utils";
+import {
+  teddsaKeygenToHex,
+  runEd25519KeygenAndSplit,
+} from "@oko-wallet-attached/crypto/keygen_ed25519";
+
+/**
+ * Handle existing user who has secp256k1 wallet but needs ed25519 keygen.
+ * Called when checkEmailV2 returns CheckEmailResponseV2NeedsEd25519Keygen.
+ *
+ * Flow: ed25519 keygen first -> get secp256k1 public_key from response -> combine secp256k1 shares
+ */
+export async function handleExistingUserNeedsEd25519Keygen(
+  idToken: string,
+  keyshareNodeMetaSecp256k1: KeyShareNodeMetaWithNodeStatusInfo,
+  keyshareNodeMetaEd25519: KeyShareNodeMetaWithNodeStatusInfo,
+  authType: AuthType,
+): Promise<Result<UserSignInResultV2, OAuthSignInError>> {
+  // 1. ed25519 keygen and split
+  const ed25519KeygenSplitRes = await runEd25519KeygenAndSplit(
+    keyshareNodeMetaEd25519,
+  );
+  if (ed25519KeygenSplitRes.success === false) {
+    return { success: false, err: ed25519KeygenSplitRes.err };
+  }
+  const {
+    keygen1: ed25519Keygen1,
+    keygen2: ed25519Keygen2,
+    userKeyShares: ed25519UserKeyShares,
+  } = ed25519KeygenSplitRes.data;
+
+  // 2. Commit to oko_api and ks nodes
+  const ksnCommitTargets: KsnCommitTarget[] = keyshareNodeMetaEd25519.nodes.map(
+    (node) => ({
+      nodeUrl: node.endpoint,
+      operationType: "add_ed25519",
+    }),
+  );
+  const commitRes = await commitAll(
+    "add_ed25519",
+    authType,
+    idToken,
+    ksnCommitTargets,
+  );
+  if (!commitRes.success) {
+    return {
+      success: false,
+      err: { type: "sign_in_request_fail", error: commitRes.err },
+    };
+  }
+  const session = commitRes.data;
+
+  // 3. Send ed25519 key shares to ks nodes using registerKeyShareEd25519V2
+  const registerEd25519Results: Result<void, string>[] = await Promise.all(
+    keyshareNodeMetaEd25519.nodes.map(async (node, index) => {
+      const commitRevealRes = createKsnCommitRevealParams(
+        session,
+        node.endpoint,
+        "register_ed25519",
+      );
+      if (!commitRevealRes.success) {
+        return { success: false, err: commitRevealRes.err };
+      }
+      return registerKeyShareEd25519V2(
+        node.endpoint,
+        idToken,
+        authType,
+        ed25519Keygen1.public_key.toHex(),
+        teddsaKeyShareToHex(ed25519UserKeyShares[index].share),
+        commitRevealRes.data,
+      );
+    }),
+  );
+  const registerEd25519ErrResults = registerEd25519Results.filter(
+    (result) => result.success === false,
+  );
+  if (registerEd25519ErrResults.length > 0) {
+    return {
+      success: false,
+      err: {
+        type: "sign_in_request_fail",
+        error: registerEd25519ErrResults.map((result) => result.err).join("\n"),
+      },
+    };
+  }
+
+  // 4. Call keygenEd25519 Oko API
+  // Use cr_final=false because secp256k1 reshare might be needed
+  const keygenEd25519CommitRevealRes = createOkoApiCommitRevealParams(
+    session,
+    "keygen_ed25519",
+    false, // cr_final: false - secp256k1 reshare might come after
+  );
+  if (!keygenEd25519CommitRevealRes.success) {
+    return {
+      success: false,
+      err: {
+        type: "sign_in_request_fail",
+        error: keygenEd25519CommitRevealRes.err,
+      },
+    };
+  }
+  const reqKeygenEd25519Res = await reqKeygenEd25519(
+    TSS_V2_ENDPOINT,
+    {
+      auth_type: authType,
+      keygen_2: {
+        key_package: serializeKeyPackage(ed25519Keygen2.key_package),
+        public_key_package: serializePublicKeyPackage(
+          ed25519Keygen2.public_key_package,
+        ),
+        identifier: [...ed25519Keygen2.identifier],
+        public_key: [...ed25519Keygen2.public_key.toUint8Array()],
+      },
+    },
+    idToken,
+    keygenEd25519CommitRevealRes.data,
+  );
+  if (reqKeygenEd25519Res.success === false) {
+    return {
+      success: false,
+      err: { type: "sign_in_request_fail", error: reqKeygenEd25519Res.msg },
+    };
+  }
+
+  // 5. Get secp256k1 public key from keygenEd25519 response
+  const secp256k1PublicKey = reqKeygenEd25519Res.data.user.public_key_secp256k1;
+
+  // 6. Request secp256k1 shares from ks nodes
+  // Use continueOnWalletNotFound=true to support auto-reshare
+  const requestSharesRes = await requestKeySharesV2WithReshareInfo(
+    idToken,
+    keyshareNodeMetaSecp256k1.nodes,
+    keyshareNodeMetaSecp256k1.threshold,
+    authType,
+    {
+      secp256k1: secp256k1PublicKey,
+    },
+    session,
+    false, // isFinal: false - reshare might come after
+    true, // continueOnWalletNotFound: true for auto-reshare
+  );
+  if (!requestSharesRes.success) {
+    const error = requestSharesRes.err;
+    console.error(
+      `[attached] insufficient shares: got ${error.got}/${error.need}`,
+    );
+    return {
+      success: false,
+      err: {
+        type: "insufficient_shares",
+      },
+    };
+  }
+
+  const { shares: keySharesByNode, nodesNeedingReshare } = requestSharesRes.data;
+  const needsReshare = nodesNeedingReshare.length > 0;
+
+  if (needsReshare) {
+    console.log(
+      "[attached] auto-reshare: detected %d nodes needing secp256k1 reshare",
+      nodesNeedingReshare.length,
+    );
+  }
+
+  // 7. Decode secp256k1 shares
+  const secp256k1DecodeRes = await decodeSecp256k1SharesByNode(keySharesByNode);
+  if (!secp256k1DecodeRes.success) {
+    return { success: false, err: secp256k1DecodeRes.err };
+  }
+
+  // 8. Auto-reshare secp256k1 if needed, or just combine
+  let keyshare1Secp256k1: string;
+
+  if (needsReshare) {
+    console.log(
+      "[attached] auto-reshare: expanding and sending secp256k1 to %d nodes",
+      nodesNeedingReshare.length,
+    );
+
+    const reshareRes = await expandAndSendReshareV2({
+      idToken,
+      authType,
+      session,
+      nodesNeedingReshare,
+      secp256k1: {
+        shares: secp256k1DecodeRes.data,
+        threshold: keyshareNodeMetaSecp256k1.threshold,
+        publicKey: secp256k1PublicKey,
+      },
+      // ed25519 is not provided - already registered via register_ed25519
+    });
+    if (!reshareRes.success) {
+      return {
+        success: false,
+        err: { type: "reshare_fail", error: reshareRes.err },
+      };
+    }
+    // keyshare1Secp256k1 is guaranteed to exist when secp256k1 is provided
+    keyshare1Secp256k1 = reshareRes.data.keyshare1Secp256k1!;
+
+    console.log("[attached] auto-reshare secp256k1 completed successfully");
+  } else {
+    // No reshare needed - just combine shares
+    const keyshare1Secp256k1Res = await combineUserShares(
+      secp256k1DecodeRes.data,
+      keyshareNodeMetaSecp256k1.threshold,
+    );
+    if (keyshare1Secp256k1Res.success === false) {
+      return {
+        success: false,
+        err: {
+          type: "key_share_combine_fail",
+          error: `secp256k1 combine err: ${keyshare1Secp256k1Res.err}`,
+        },
+      };
+    }
+    keyshare1Secp256k1 = keyshare1Secp256k1Res.data;
+  }
+
+  // 9. Convert ed25519 keygen1 to hex format for storage
+  const keyPackageEd25519Hex = teddsaKeygenToHex(ed25519Keygen1);
+
+  return {
+    success: true,
+    data: {
+      publicKeySecp256k1: reqKeygenEd25519Res.data.user.public_key_secp256k1,
+      publicKeyEd25519: reqKeygenEd25519Res.data.user.public_key_ed25519,
+      walletIdSecp256k1: reqKeygenEd25519Res.data.user.wallet_id_secp256k1,
+      walletIdEd25519: reqKeygenEd25519Res.data.user.wallet_id_ed25519,
+      jwtToken: reqKeygenEd25519Res.data.token,
+      keyshare1Secp256k1,
+      keyPackageEd25519: keyPackageEd25519Hex.keyPackage,
+      publicKeyPackageEd25519: keyPackageEd25519Hex.publicKeyPackage,
+      isNewUser: false,
+      email: reqKeygenEd25519Res.data.user.email ?? null,
+      name: reqKeygenEd25519Res.data.user.name ?? null,
+    },
+  };
+}
+
+/**
+ * Handle reshare for secp256k1 + keygen for ed25519 (scenario 6).
+ * Called when user has secp256k1 wallet, needs reshare, and needs ed25519 keygen.
+ *
+ * Commit-reveal:
+ * - oko_api: sign_in_reshare_ed25519
+ * - KSN ACTIVE nodes: sign_in_reshare_ed25519
+ * - KSN new nodes: register_reshare
+ */
+export async function handleReshareAndEd25519Keygen(
+  idToken: string,
+  keyshareNodeMetaSecp256k1: KeyShareNodeMetaWithNodeStatusInfo,
+  keyshareNodeMetaEd25519: KeyShareNodeMetaWithNodeStatusInfo,
+  authType: AuthType,
+): Promise<Result<UserSignInResultV2, OAuthSignInError>> {
+  // 1. Classify nodes
+  const activeNodes = keyshareNodeMetaSecp256k1.nodes.filter(
+    (n) => n.wallet_status === "ACTIVE",
+  );
+  const newNodes = keyshareNodeMetaSecp256k1.nodes.filter(
+    (n) =>
+      n.wallet_status === "NOT_REGISTERED" ||
+      n.wallet_status === "UNRECOVERABLE_DATA_LOSS",
+  );
+
+  if (activeNodes.length < keyshareNodeMetaSecp256k1.threshold) {
+    return {
+      success: false,
+      err: {
+        type: "reshare_fail",
+        error: "insufficient existing KS nodes for reshare",
+      },
+    };
+  }
+
+  // 2. ed25519 keygen and split
+  const ed25519KeygenSplitRes = await runEd25519KeygenAndSplit(
+    keyshareNodeMetaEd25519,
+  );
+  if (ed25519KeygenSplitRes.success === false) {
+    return { success: false, err: ed25519KeygenSplitRes.err };
+  }
+  const {
+    keygen1: ed25519Keygen1,
+    keygen2: ed25519Keygen2,
+    userKeyShares: ed25519UserKeyShares,
+  } = ed25519KeygenSplitRes.data;
+
+  // 3. Commit to oko_api and ks nodes
+  const ksnCommitTargets: KsnCommitTarget[] = [
+    ...activeNodes.map((node) => ({
+      nodeUrl: node.endpoint,
+      operationType: "add_ed25519" as const,
+    })),
+    ...newNodes.map((node) => ({
+      nodeUrl: node.endpoint,
+      operationType: "add_ed25519" as const,
+    })),
+  ];
+  const commitRes = await commitAll(
+    "add_ed25519",
+    authType,
+    idToken,
+    ksnCommitTargets,
+  );
+  if (!commitRes.success) {
+    return {
+      success: false,
+      err: { type: "reshare_fail", error: commitRes.err },
+    };
+  }
+  const session = commitRes.data;
+
+  // 4. Sign in to get the public key
+  const signInCommitRevealRes = createOkoApiCommitRevealParams(
+    session,
+    "signin",
+  );
+  if (!signInCommitRevealRes.success) {
+    return {
+      success: false,
+      err: { type: "reshare_fail", error: signInCommitRevealRes.err },
+    };
+  }
+  const signInResult = await signInV2(
+    idToken,
+    authType,
+    signInCommitRevealRes.data,
+  );
+  if (!signInResult.success) {
+    return { success: false, err: signInResult.err };
+  }
+  const signInResp = signInResult.data;
+  const secp256k1PublicKey = signInResp.user.public_key_secp256k1;
+
+  // 5. Request secp256k1 shares with public key
+  const requestSecp256k1SharesRes = await requestKeySharesV2(
+    idToken,
+    activeNodes,
+    keyshareNodeMetaSecp256k1.threshold,
+    authType,
+    {
+      secp256k1: secp256k1PublicKey,
+    },
+    session,
+  );
+  if (!requestSecp256k1SharesRes.success) {
+    return {
+      success: false,
+      err: {
+        type: "reshare_fail",
+        error: `Failed to request secp256k1 shares: ${requestSecp256k1SharesRes.err.code}`,
+      },
+    };
+  }
+
+  // 6. Decode secp256k1 shares (filter out items without shares for reshare case)
+  const itemsWithSecp256k1 = requestSecp256k1SharesRes.data.filter(
+    (item) => item.shares.secp256k1,
+  );
+  const secp256k1DecodeRes =
+    await decodeSecp256k1SharesByNode(itemsWithSecp256k1);
+  if (!secp256k1DecodeRes.success) {
+    return {
+      success: false,
+      err: {
+        type: "reshare_fail",
+        error: secp256k1DecodeRes.err.error,
+      },
+    };
+  }
+
+  // 7. Expand secp256k1 shares to new nodes
+  const secp256k1ExpandRes = await runExpandShares(
+    secp256k1DecodeRes.data,
+    newNodes,
+    keyshareNodeMetaSecp256k1.threshold,
+  );
+  if (!secp256k1ExpandRes.success) {
+    return {
+      success: false,
+      err: { type: "reshare_fail", error: secp256k1ExpandRes.err },
+    };
+  }
+
+  // 8. Send shares to KSN
+  // For ACTIVE nodes: reshare secp256k1 + register ed25519
+  // For new nodes: reshare/register both
+  const allNodes = keyshareNodeMetaSecp256k1.nodes;
+  const sendResults = await Promise.all(
+    secp256k1ExpandRes.data.reshared_user_key_shares.map(
+      async (secp256k1Share) => {
+        const ed25519Share = ed25519UserKeyShares.find(
+          (s) => s.node.endpoint === secp256k1Share.node.endpoint,
+        );
+        if (!ed25519Share) {
+          return { success: false, err: "ed25519 share not found for node" };
+        }
+
+        const node = secp256k1Share.node;
+        const nodeStatus = allNodes.find(
+          (n) => n.endpoint === node.endpoint,
+        )?.wallet_status;
+
+        const isNewNode =
+          nodeStatus === "NOT_REGISTERED" ||
+          nodeStatus === "UNRECOVERABLE_DATA_LOSS";
+
+        if (isNewNode) {
+          // New node: reshare/register both curves
+          const commitRevealRes = createKsnCommitRevealParams(
+            session,
+            node.endpoint,
+            "reshare_register",
+            true, // cr_final: true - reshare_register is the final KSN call for new nodes
+          );
+          if (!commitRevealRes.success) {
+            return { success: false, err: commitRevealRes.err };
+          }
+          return reshareRegisterV2(
+            node.endpoint,
+            idToken,
+            authType,
+            {
+              secp256k1: {
+                public_key: secp256k1PublicKey,
+                share: encodePoint256ToKeyShareString(secp256k1Share.share),
+              },
+              ed25519: {
+                public_key: ed25519Keygen1.public_key.toHex(),
+                share: teddsaKeyShareToHex(ed25519Share.share),
+              },
+            },
+            commitRevealRes.data,
+          );
+        } else {
+          // ACTIVE node: reshare secp256k1 + register ed25519 separately
+          const reshareCommitRevealRes = createKsnCommitRevealParams(
+            session,
+            node.endpoint,
+            "reshare",
+          );
+          if (!reshareCommitRevealRes.success) {
+            return { success: false, err: reshareCommitRevealRes.err };
+          }
+          const reshareSecp256k1Res = await reshareKeySharesV2(
+            node.endpoint,
+            idToken,
+            authType,
+            {
+              secp256k1: {
+                public_key: secp256k1PublicKey,
+                share: encodePoint256ToKeyShareString(secp256k1Share.share),
+              },
+            },
+            reshareCommitRevealRes.data,
+          );
+          if (!reshareSecp256k1Res.success) {
+            return reshareSecp256k1Res;
+          }
+
+          const registerEd25519CommitRevealRes = createKsnCommitRevealParams(
+            session,
+            node.endpoint,
+            "register_ed25519",
+            true, // cr_final: true - register_ed25519 is the final KSN call for ACTIVE nodes
+          );
+          if (!registerEd25519CommitRevealRes.success) {
+            return { success: false, err: registerEd25519CommitRevealRes.err };
+          }
+          return registerKeyShareEd25519V2(
+            node.endpoint,
+            idToken,
+            authType,
+            ed25519Keygen1.public_key.toHex(),
+            teddsaKeyShareToHex(ed25519Share.share),
+            registerEd25519CommitRevealRes.data,
+          );
+        }
+      },
+    ),
+  );
+
+  const errResults = sendResults.filter((r) => !r.success);
+  if (errResults.length > 0) {
+    return {
+      success: false,
+      err: {
+        type: "reshare_fail",
+        error: errResults.map((r) => (r as { err: string }).err).join("\n"),
+      },
+    };
+  }
+
+  // 9. Update Oko API reshare status (not final - keygen_ed25519 comes after)
+  const reshareCommitRevealRes = createOkoApiCommitRevealParams(
+    session,
+    "reshare",
+    false, // cr_final: false - keygen_ed25519 is the final oko_api call
+  );
+  if (!reshareCommitRevealRes.success) {
+    return {
+      success: false,
+      err: { type: "reshare_fail", error: reshareCommitRevealRes.err },
+    };
+  }
+  const resharedNodes = secp256k1ExpandRes.data.reshared_user_key_shares.map(
+    (s) => s.node,
+  );
+  const updateRes = await makeAuthorizedOkoApiRequest<ReshareRequestV2, void>(
+    "user/reshare",
+    idToken,
+    {
+      wallets: {
+        secp256k1: {
+          public_key: secp256k1PublicKey,
+          reshared_key_shares: resharedNodes,
+        },
+      },
+    },
+    TSS_V2_ENDPOINT,
+    reshareCommitRevealRes.data,
+  );
+  if (!updateRes.success) {
+    console.warn("[attached] Failed to update reshare status:", updateRes.err);
+  }
+
+  // 10. Call keygenEd25519 API (FINAL - must be last oko_api call)
+  const keygenEd25519CommitRevealRes = createOkoApiCommitRevealParams(
+    session,
+    "keygen_ed25519",
+    true, // cr_final: true - keygen_ed25519 is the final oko_api call
+  );
+  if (!keygenEd25519CommitRevealRes.success) {
+    return {
+      success: false,
+      err: { type: "reshare_fail", error: keygenEd25519CommitRevealRes.err },
+    };
+  }
+  const reqKeygenEd25519Res = await reqKeygenEd25519(
+    TSS_V2_ENDPOINT,
+    {
+      auth_type: authType,
+      keygen_2: {
+        key_package: serializeKeyPackage(ed25519Keygen2.key_package),
+        public_key_package: serializePublicKeyPackage(
+          ed25519Keygen2.public_key_package,
+        ),
+        identifier: [...ed25519Keygen2.identifier],
+        public_key: [...ed25519Keygen2.public_key.toUint8Array()],
+      },
+    },
+    idToken,
+    keygenEd25519CommitRevealRes.data,
+  );
+  if (reqKeygenEd25519Res.success === false) {
+    return {
+      success: false,
+      err: { type: "sign_in_request_fail", error: reqKeygenEd25519Res.msg },
+    };
+  }
+
+  // 11. Convert ed25519 keygen1 to hex format for storage
+  const keyPackageEd25519Hex = teddsaKeygenToHex(ed25519Keygen1);
+
+  return {
+    success: true,
+    data: {
+      publicKeySecp256k1: reqKeygenEd25519Res.data.user.public_key_secp256k1,
+      publicKeyEd25519: reqKeygenEd25519Res.data.user.public_key_ed25519,
+      walletIdSecp256k1: reqKeygenEd25519Res.data.user.wallet_id_secp256k1,
+      walletIdEd25519: reqKeygenEd25519Res.data.user.wallet_id_ed25519,
+      jwtToken: reqKeygenEd25519Res.data.token,
+      keyshare1Secp256k1: secp256k1ExpandRes.data.original_secret.toHex(),
+      keyPackageEd25519: keyPackageEd25519Hex.keyPackage,
+      publicKeyPackageEd25519: keyPackageEd25519Hex.publicKeyPackage,
+      isNewUser: false,
+      email: reqKeygenEd25519Res.data.user.email ?? null,
+      name: reqKeygenEd25519Res.data.user.name ?? null,
+    },
+  };
+}
